Added main activity UI and functionality along with validation

+++ app/src/main/java/raghu/co/repository/network/RestApi.java
@@ -2,7 +2,6 @@ package raghu.co.repository.network;
 
 import io.reactivex.Flowable;
 import raghu.co.repository.model.cars.RentalCars;
-import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 
@@ -11,8 +10,8 @@ public interface RestApi {
     @GET("v1.2/cars/search-circle")
     Flowable<RentalCars> getRentalCars(
             @Query("apikey") String apiKey,
-            @Query("latitude") float latitude,
-            @Query("longitude") float longitude,
+            @Query("latitude") double latitude,
+            @Query("longitude") double longitude,
             @Query("radius") int radius,
             @Query("pick_up") String pick_up,
             @Query("drop_off") String drop_off
